requirements.txt
streamlit==1.32.0
pandas==2.0.3
plotly==5.18.0
Pillow==10.1.0
python-dotenv==1.0.0